[refactor] build: split composeApp into KMP library + androidApp module
AGP 9 deprecates combining com.android.application with the Kotlin
Multiplatform plugin on the same module, which is also why
kotlinter-gradle >=5.4.0 double-registers its lintKotlin task on
composeApp (matches both org.jetbrains.kotlin.multiplatform and
com.android.base in its plugin-detection map), breaking CI on the
renovate/kotlinter branch.

Per Kotlin's AGP 9 migration guide: composeApp now applies
com.android.kotlin.multiplatform.library instead and becomes a pure
KMP library (wasmJs, android-as-library, desktop, iOS unchanged). The
Android application (MainActivity, manifest, signing, ProGuard, build
variants, managed-device instrumented tests) moves to a new plain
androidApp module that depends on composeApp.

Also bumps kotlinter 5.3.0 -> 5.7.0, the change that originally
surfaced this, now that the two plugins are no longer co-applied on
one module. Updates CI task/output paths and docs to match.
