+++ .github/workflows/tag_create_release.yml
@@ -76,13 +76,13 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.PAT_FOR_RELEASES }}
         with:
           upload_url: ${{ steps.create_release.outputs.upload_url }}
-          asset_path: ./composeApp/build/outputs/mapping/release/mapping.txt
+          asset_path: ./androidApp/build/outputs/mapping/release/mapping.txt
           asset_name: mapping-${{ env.version }}.txt
           asset_content_type: text/plain
 
       - name: Find APK file
         run: |
-          apk_path=$(find ./composeApp/build/outputs/apk/release -name "${{ env.app_name }}-*.apk" | head -n 1)
+          apk_path=$(find ./androidApp/build/outputs/apk/release -name "${{ env.app_name }}-*.apk" | head -n 1)
           echo "apk_path=$apk_path" >> $GITHUB_ENV
 
       - name: Upload Release Asset APK
@@ -97,7 +97,7 @@ jobs:
 
       - name: Find AAB file
         run: |
-          aab_path=$(find ./composeApp/build/outputs/bundle/release -name "${{ env.app_name }}-*.aab" | head -n 1)
+          aab_path=$(find ./androidApp/build/outputs/bundle/release -name "${{ env.app_name }}-*.aab" | head -n 1)
           echo "aab_path=$aab_path" >> $GITHUB_ENV
 
       - name: Upload Release Asset AAB
