+++ composeApp/build.gradle.kts
@@ -1,24 +1,17 @@
 @file:OptIn(ExperimentalWasmDsl::class)
 
-import com.android.build.api.dsl.ManagedVirtualDevice
-import com.android.build.gradle.internal.dsl.BaseAppModuleExtension
-import java.io.FileInputStream
-import java.io.InputStreamReader
-import java.text.SimpleDateFormat
-import java.util.Date
-import java.util.Properties
 import org.jetbrains.compose.desktop.application.dsl.TargetFormat
 import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
 import org.jetbrains.kotlin.gradle.ExperimentalWasmDsl
 import org.jetbrains.kotlin.gradle.dsl.JvmTarget
 import org.jetbrains.kotlin.gradle.targets.js.webpack.KotlinWebpackConfig
+import org.jetbrains.kotlin.gradle.tasks.KotlinCompileTool
 import org.jmailen.gradle.kotlinter.tasks.FormatTask
 import org.jmailen.gradle.kotlinter.tasks.LintTask
 
 plugins {
-    id("base")
     alias(libs.plugins.kotlinMultiplatform)
-    alias(libs.plugins.androidApplication)
+    alias(libs.plugins.androidKotlinMultiplatformLibrary)
     alias(libs.plugins.jetbrainsCompose)
     alias(libs.plugins.compose.compiler)
     alias(libs.plugins.detekt)
@@ -26,9 +19,7 @@ plugins {
 }
 
 // Configuration
-val productApkName = "DVDMultiplatform"
 val productNamespace = "com.rwmobi.dvdmultiplatform"
-val isRunningOnCI = System.getenv("CI") == "true"
 
 kotlin {
     wasmJs {
@@ -49,7 +40,18 @@ kotlin {
         binaries.executable()
     }
 
-    androidTarget {
+    android {
+        // Must differ from androidApp's namespace/applicationId — AGP rejects two modules
+        // sharing one namespace once composeApp and androidApp are separate modules.
+        namespace = "$productNamespace.shared"
+        compileSdk = libs.versions.compileSdk.get().toInt()
+        minSdk = libs.versions.minSdk.get().toInt()
+
+        // Required so Compose Multiplatform's composeResources (drawables, strings, etc.) are
+        // packaged as real Android resources — otherwise androidApp gets MissingResourceException
+        // at runtime, since it only depends on composeApp as a library, not its raw source set.
+        androidResources.enable = true
+
         @OptIn(ExperimentalKotlinGradlePluginApi::class)
         compilerOptions {
             jvmTarget.set(JvmTarget.JVM_17)
@@ -72,17 +74,6 @@ kotlin {
     sourceSets {
         val desktopMain by getting
 
-        androidMain.dependencies {
-            implementation(compose.preview)
-            implementation(libs.androidx.activity.compose)
-        }
-        androidInstrumentedTest.dependencies {
-            implementation(compose.ui)
-            implementation(libs.androidx.test.ext.junit)
-            implementation(libs.espresso.core)
-            implementation(libs.ui.test.junit4)
-            implementation(libs.androidx.test.rules)
-        }
         commonMain.dependencies {
             implementation(compose.runtime)
             implementation(compose.foundation)
@@ -97,56 +88,10 @@ kotlin {
     }
 }
 
-android {
-    namespace = productNamespace
-
-    setupSdkVersionsFromVersionCatalog()
-    setupSigningAndBuildTypes()
-    setupPackagingResourcesDeduplication()
-
-    sourceSets["main"].manifest.srcFile("src/androidMain/AndroidManifest.xml")
-    sourceSets["main"].res.srcDirs("src/androidMain/res")
-    sourceSets["main"].resources.srcDirs("src/commonMain/resources")
-
-    defaultConfig {
-        applicationId = productNamespace
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        vectorDrawables { useSupportLibrary = true }
-    }
-
-    buildFeatures {
-        compose = true
-        buildConfig = true
-    }
-
-    testOptions {
-        animationsDisabled = true
-
-        unitTests {
-            isIncludeAndroidResources = true
-            isReturnDefaultValues = true
-        }
-
-        managedDevices {
-            allDevices {
-                create<ManagedVirtualDevice>("pixel2Api35") {
-                    device = "Pixel 2"
-                    apiLevel = 35
-                    systemImageSource = "aosp-atd"
-                }
-            }
-        }
-    }
-
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_17
-        targetCompatibility = JavaVersion.VERSION_17
-    }
-
-    dependencies {
-        debugImplementation(compose.uiTooling)
-    }
+compose.resources {
+    // Generated Res accessors are internal by default; androidApp (a separate module since the
+    // KMP-android-library split) needs to reach them from its instrumented tests.
+    publicResClass = true
 }
 
 compose.desktop {
@@ -163,8 +108,11 @@ compose.desktop {
 
 tasks {
     check { dependsOn("detekt") }
-    preBuild { dependsOn("formatKotlin") }
 }
+
+// The KMP android library plugin doesn't create a 'preBuild' lifecycle task like
+// com.android.application/library did, so hook formatting into every Kotlin compile task instead.
+tasks.withType<KotlinCompileTool>().configureEach { dependsOn("formatKotlin") }
 
 tasks.withType<LintTask> {
     exclude {
@@ -178,117 +126,3 @@ tasks.withType<LintTask> {
 tasks.withType<FormatTask> { exclude { it.file.path.contains("generated/") } }
 
 detekt { parallel = true }
-
-// Gradle Build Utilities
-private fun BaseAppModuleExtension.setupSdkVersionsFromVersionCatalog() {
-    compileSdk = libs.versions.compileSdk.get().toInt()
-    defaultConfig {
-        minSdk = libs.versions.minSdk.get().toInt()
-        targetSdk = libs.versions.targetSdk.get().toInt()
-        versionCode = libs.versions.versionCode.get().toInt()
-        versionName = libs.versions.versionName.get()
-    }
-}
-
-private fun BaseAppModuleExtension.setupPackagingResourcesDeduplication() {
-    packaging.resources {
-        excludes.addAll(
-            listOf(
-                "META-INF/*.md",
-                "META-INF/proguard/*",
-                "META-INF/*.kotlin_module",
-                "META-INF/DEPENDENCIES",
-                "META-INF/LICENSE",
-                "META-INF/LICENSE.*",
-                "META-INF/LICENSE-notice.txt",
-                "META-INF/NOTICE",
-                "META-INF/NOTICE.*",
-                "META-INF/AL2.0",
-                "META-INF/LGPL2.1",
-                "META-INF/*.properties",
-                "/*.properties",
-            ),
-        )
-    }
-}
-
-private fun BaseAppModuleExtension.setupSigningAndBuildTypes() {
-    val releaseSigningConfigName = "releaseSigningConfig"
-    val timestamp = SimpleDateFormat("yyyyMMdd-HHmmss").format(Date())
-    val baseName = "$productApkName-${libs.versions.versionName.get()}-$timestamp"
-    val isReleaseBuild = gradle.startParameter.taskNames.any {
-        it.contains("Release", ignoreCase = true)
-                || it.contains("Bundle", ignoreCase = true)
-                || it.equals("build", ignoreCase = true)
-    }
-
-    project.extensions.configure<BasePluginExtension> { archivesName.set(baseName) }
-
-    signingConfigs.create(releaseSigningConfigName) {
-        // Only initialise the signing config when a Release or Bundle task is being executed.
-        // This prevents Gradle sync or debug builds from attempting to load the keystore,
-        // which could fail if the keystore or environment variables are not available.
-        // SigningConfig itself is only wired to the 'release' build type, so this guard avoids unnecessary setup.
-        if (isReleaseBuild) {
-            val keystorePropertiesFile = file("../../keystore.properties")
-
-            if (isRunningOnCI || !keystorePropertiesFile.exists()) {
-                println("⚠\uFE0F Signing Config: using environment variables")
-                keyAlias = System.getenv("CI_ANDROID_KEYSTORE_ALIAS")
-                keyPassword = System.getenv("CI_ANDROID_KEYSTORE_PRIVATE_KEY_PASSWORD")
-                storeFile = file(System.getenv("KEYSTORE_LOCATION"))
-                storePassword = System.getenv("CI_ANDROID_KEYSTORE_PASSWORD")
-            } else {
-                println("⚠\uFE0F Signing Config: using keystore properties")
-                val properties = Properties()
-                InputStreamReader(
-                    FileInputStream(keystorePropertiesFile),
-                    Charsets.UTF_8,
-                ).use { reader ->
-                    properties.load(reader)
-                }
-
-                keyAlias = properties.getProperty("alias")
-                keyPassword = properties.getProperty("pass")
-                storeFile = file(properties.getProperty("store"))
-                storePassword = properties.getProperty("storePass")
-            }
-        } else {
-            println("⚠\uFE0F Signing Config: not created for non-release builds.")
-        }
-    }
-
-    buildTypes {
-        fun setOutputFileName() {
-            applicationVariants.all {
-                outputs
-                    .map { it as com.android.build.gradle.internal.api.BaseVariantOutputImpl }
-                    .forEach { output ->
-                        val outputFileName = "$productApkName-$name-$versionName-$timestamp.apk"
-                        output.outputFileName = outputFileName
-                    }
-            }
-        }
-
-        getByName("debug") {
-            applicationIdSuffix = ".debug"
-            isMinifyEnabled = false
-            isDebuggable = true
-            setOutputFileName()
-        }
-
-        getByName("release") {
-            isShrinkResources = true
-            isMinifyEnabled = true
-            isDebuggable = false
-            setProguardFiles(
-                listOf(
-                    getDefaultProguardFile("proguard-android-optimize.txt"),
-                    "proguard-rules.pro",
-                ),
-            )
-            signingConfig = signingConfigs.getByName(name = releaseSigningConfigName)
-            setOutputFileName()
-        }
-    }
-}
