+++ gradle/libs.versions.toml
@@ -7,7 +7,7 @@ espresso-core = "3.7.0"
 testRules = "1.7.0"
 androidx-test-ext-junit = "1.3.0"
 compose-bom = "2026.04.01"
-kotlinter = "5.3.0"
+kotlinter = "5.7.0"
 detekt = "1.23.8"
 
 # App configurations, not dependencies
@@ -28,6 +28,8 @@ androidx-test-ext-junit = { group = "androidx.test.ext", name = "junit", version
 [plugins]
 androidApplication = { id = "com.android.application", version.ref = "agp" }
 androidLibrary = { id = "com.android.library", version.ref = "agp" }
+androidKotlinMultiplatformLibrary = { id = "com.android.kotlin.multiplatform.library", version.ref = "agp" }
+kotlinAndroid = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin" }
 jetbrainsCompose = { id = "org.jetbrains.compose", version.ref = "compose-plugin" }
 compose-compiler = { id = "org.jetbrains.kotlin.plugin.compose", version.ref = "kotlin" }
 kotlinMultiplatform = { id = "org.jetbrains.kotlin.multiplatform", version.ref = "kotlin" }
